requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
opencv-python
numpy
flask
pillow
requests==2.22.0
flask-cors
gunicorn
torch==1.9.0
torchvision==0.10.0
#'git+https://github.com/facebookresearch/detectron2.git'



